+++ src/commands/borealis-pg/extensions/index.ts
@@ -1,19 +1,17 @@
 import {HTTP, HTTPError} from '@heroku/http-call'
 import color from '@heroku-cli/color'
 import {Command} from '@heroku-cli/command'
+import Table, {Header} from 'tty-table'
 import {applyActionSpinner} from '../../../async-actions'
 import {getBorealisPgApiUrl, getBorealisPgAuthHeader} from '../../../borealis-api'
 import {
   addonOptionName,
   appOptionName,
+  baseTableHeaderStyle,
   cliOptions,
-  consoleColours,
   processAddonAttachmentInfo,
 } from '../../../command-components'
 import {createHerokuAuth, fetchAddonAttachmentInfo, removeHerokuAuth} from '../../../heroku-api'
-
-const pgExtensionColour = consoleColours.pgExtension
-const pgExtMetadataColour = consoleColours.dataFieldValue
 
 export default class ListPgExtensionsCommand extends Command {
   static description =
@@ -37,22 +35,14 @@ export default class ListPgExtensionsCommand extends Command {
     try {
       const response = await applyActionSpinner(
         `Fetching Postgres extension list for add-on ${color.addon(addonName)}`,
-        HTTP.get(getBorealisPgApiUrl(`/heroku/resources/${addonName}/pg-extensions`), {
-          headers: {Authorization: getBorealisPgAuthHeader(authorization)},
-        }),
+        HTTP.get<{extensions: ExtensionInfo[]}>(
+          getBorealisPgApiUrl(`/heroku/resources/${addonName}/pg-extensions`),
+          {headers: {Authorization: getBorealisPgAuthHeader(authorization)}},
+        ),
       )
 
-      const responseBody = response.body as {
-        extensions: Array<{name: string; schema: string; version: string}>
-      }
-      if (responseBody.extensions.length > 0) {
-        for (const extInfo of responseBody.extensions) {
-          this.log(
-            `- ${pgExtensionColour(extInfo.name)} ` +
-              `(version: ${pgExtMetadataColour(extInfo.version)}, ` +
-              `schema: ${pgExtMetadataColour(extInfo.schema)})`,
-          )
-        }
+      if (response.body.extensions.length > 0) {
+        this.log(renderResultsTable(response.body.extensions))
       } else {
         this.warn('No extensions found')
       }
@@ -75,4 +65,22 @@ export default class ListPgExtensionsCommand extends Command {
       throw err
     }
   }
+}
+
+function renderResultsTable(extensions: ExtensionInfo[]) {
+  const headers: Header[] = [
+    {...baseTableHeaderStyle, alias: 'Name', value: 'name'},
+    {...baseTableHeaderStyle, alias: 'Version', value: 'version'},
+    {...baseTableHeaderStyle, alias: 'Schema', value: 'schema'},
+  ]
+
+  const table = Table(headers, extensions, {truncate: false, borderStyle: 'dashed', compact: true})
+
+  return table.render()
+}
+
+interface ExtensionInfo {
+  name: string
+  schema: string
+  version: string
 }